Added a bonus for Archers against Musketeers and Dragoons.
I was thinking for a long time that I'd like to make archers relevant
against early muskets to an extent, similar to the real world. Now
I've done that by adding bonuses for archers against those units.
Archers can now fight on even terms with Musketeers and Dragoons
(but are otherwise the same as before).